moved utils to utils directory
