+++ packages/playground/tests/e2e/cards/utils/cardUtils.js
@@ -1,6 +1,6 @@
-import { setIframeSelector, fillIFrame, deleteFromIFrame } from '../commonUtils';
+import { setIframeSelector, fillIFrame, deleteFromIFrame } from '../../commonUtils';
 
-import { REGULAR_TEST_CARD, TEST_DATE_VALUE, TEST_CVC_VALUE } from '../constants';
+import { REGULAR_TEST_CARD, TEST_DATE_VALUE, TEST_CVC_VALUE } from '../../constants';
 
 /**
  * Unique to each component are where the iframes are to be found,
